Moved SWT authorization to its own package to remove extra dependencies
from Commands and keep them separated
